Fixed compiler warning.

+++ copasi/model/CReaction.cpp
@@ -1,9 +1,9 @@
 // Begin CVS Header
 //   $Source: /Volumes/Home/Users/shoops/cvs/copasi_dev/copasi/model/CReaction.cpp,v $
-//   $Revision: 1.192.2.3 $
+//   $Revision: 1.192.2.4 $
 //   $Name:  $
 //   $Author: shoops $
-//   $Date: 2011/02/18 19:58:14 $
+//   $Date: 2011/03/30 12:24:45 $
 // End CVS Header
 
 // Copyright (C) 2011 - 2010 by Pedro Mendes, Virginia Tech Intellectual
@@ -748,7 +748,7 @@ C_INT32 CReaction::loadOld(CReadConfig & configbuffer)
 
   // Construct parameters
   if (mMap.getFunctionParameters().getNumberOfParametersByUsage(CFunctionParameter::PARAMETER)
-      != ParameterSize)
+      != (size_t) ParameterSize)
     {
       // no. of parameters not matching function definition.
       fatalError();
@@ -759,7 +759,7 @@ C_INT32 CReaction::loadOld(CReadConfig & configbuffer)
   const CFunctionParameter* pParameter;
   C_FLOAT64 value;
 
-  for (i = 0, pos = 0; i < ParameterSize; i++)
+  for (i = 0, pos = 0; i < (size_t) ParameterSize; i++)
     {
       name = StringPrint("Param%d", i);
       configbuffer.getVariable(name, "C_FLOAT64", &value);
